requirements.txt
# -------- Core Backend --------
fastapi
uvicorn

# -------- Frontend --------
streamlit
requests

# -------- RAG / LangChain --------
langchain
langchain-community
langchain-groq

# -------- Embeddings --------
sentence-transformers
torch

# -------- Vector Database --------
chromadb

# -------- PDF Processing --------
pymupdf
pypdf

# -------- Text Splitting --------
langchain-text-splitters

# -------- Utilities --------
python-dotenv
numpy




python-multipart